requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
faiss-cpu==1.11.0
google-auth==2.39.0
google-genai==1.13.0
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
scikit-learn==1.6.1
sentence-transformers==4.1.0
streamlit==1.45.0
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
groq==0.24.0